+++ src/Type/src/Factory/DisjunctionTypeFactory.php
@@ -8,32 +8,11 @@ use spaceonfire\Type\DisjunctionType;
 use spaceonfire\Type\Exception\TypeNotSupportedException;
 use spaceonfire\Type\Type;
 
-final class DisjunctionTypeFactory implements TypeFactoryInterface
+final class DisjunctionTypeFactory extends AbstractAggregatedTypeFactory
 {
-    use TypeFactoryTrait;
-
-    /**
-     * @inheritDoc
-     */
-    public function supports(string $type): bool
+    public function __construct()
     {
-        if ($this->parent === null) {
-            return false;
-        }
-
-        $parts = explode(DisjunctionType::DELIMITER, $type);
-
-        if (count($parts) < 2) {
-            return false;
-        }
-
-        foreach ($parts as $part) {
-            if (!$this->parent->supports(trim($part))) {
-                return false;
-            }
-        }
-
-        return true;
+        parent::__construct(DisjunctionType::DELIMITER);
     }
 
     /**
@@ -41,14 +20,12 @@ final class DisjunctionTypeFactory implements TypeFactoryInterface
      */
     public function make(string $type): Type
     {
-        if (!$this->supports($type)) {
+        $parsed = $this->parse($type);
+
+        if ($parsed === null || $this->parent === null) {
             throw new TypeNotSupportedException($type, DisjunctionType::class);
         }
 
-        $conjuncts = array_map(function (string $subType): Type {
-            return $this->parent->make(trim($subType));
-        }, explode(DisjunctionType::DELIMITER, $type));
-
-        return new DisjunctionType($conjuncts);
+        return new DisjunctionType(array_map([$this->parent, 'make'], $parsed));
     }
 }