+++ src/Type/src/Factory/GroupTypeFactory.php
@@ -5,10 +5,9 @@ declare(strict_types=1);
 namespace spaceonfire\Type\Factory;
 
 use spaceonfire\Type\Exception\TypeNotSupportedException;
-use spaceonfire\Type\MixedType;
 use spaceonfire\Type\Type;
 
-final class MixedTypeFactory implements TypeFactoryInterface
+final class GroupTypeFactory implements TypeFactoryInterface
 {
     use TypeFactoryTrait;
 
@@ -17,8 +16,17 @@ final class MixedTypeFactory implements TypeFactoryInterface
      */
     public function supports(string $type): bool
     {
+        if ($this->parent === null) {
+            return false;
+        }
+
         $type = $this->removeWhitespaces($type);
-        return $type === MixedType::NAME;
+
+        return
+            strlen($type) > 2 &&
+            $type[0] === '(' &&
+            strrev($type)[0] === ')' &&
+            $this->parent->supports(substr($type, 1, -1));
     }
 
     /**
@@ -27,10 +35,11 @@ final class MixedTypeFactory implements TypeFactoryInterface
     public function make(string $type): Type
     {
         $type = $this->removeWhitespaces($type);
+
         if (!$this->supports($type)) {
-            throw new TypeNotSupportedException($type, MixedType::class);
+            throw new TypeNotSupportedException($type);
         }
 
-        return new MixedType();
+        return $this->parent->make(substr($type, 1, -1));
     }
 }