+++ src/Type/tests/CollectionTypeTest.php
@@ -35,20 +35,30 @@ class CollectionTypeTest extends TestCase
     public function testStringify(): void
     {
         $integerCollection = new CollectionType(new BuiltinType(BuiltinType::INT));
-        self::assertEquals('int[]', (string)$integerCollection);
+        self::assertSame('int[]', (string)$integerCollection);
 
         $arrayStringInt = new CollectionType(
             new BuiltinType(BuiltinType::INT),
             new BuiltinType(BuiltinType::STRING),
             new BuiltinType(BuiltinType::ARRAY)
         );
-        self::assertEquals('array<string,int>', (string)$arrayStringInt);
+        self::assertSame('array<string,int>', (string)$arrayStringInt);
 
         $arrayIteratorInteger = new CollectionType(
             new BuiltinType(BuiltinType::INT),
             null,
             new InstanceOfType(ArrayIterator::class)
         );
-        self::assertEquals('ArrayIterator<int>', (string)$arrayIteratorInteger);
+        self::assertSame('ArrayIterator<int>', (string)$arrayIteratorInteger);
+
+        $arrayOfBoolOrInt = new CollectionType(
+            new DisjunctionType([
+                new BuiltinType(BuiltinType::BOOL),
+                new BuiltinType(BuiltinType::INT),
+            ]),
+            null,
+            new BuiltinType(BuiltinType::ARRAY)
+        );
+        self::assertSame('array<bool|int>', (string)$arrayOfBoolOrInt);
     }
 }