+++ src/Type/CHANGELOG.md
@@ -22,6 +22,23 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 - Nothing
 -->
 
+## [1.3.0] - 2020-10-17
+### Added
+- Lot of improvements for `DisjunctionType` and `ConjunctionType`:
+    - General code moved to `AbstractAggregatedType` class;
+    - They can be iterated over now;
+    - They can be merged with other aggregated types using late static binding;
+    - Duplicates auto removed from given subtypes;
+    - At least 2 subtypes required.
+- `DisjunctionTypeFactory` and `ConjunctionTypeFactory` also improved:
+    - General code moved to `AbstractAggregatedTypeFactory` class;
+    - They can now parse complex type strings, such as: `int|string|array<bool|int>|string|null`.
+- `GroupTypeFactory` allows parsing grouped with `()` type strings like: `(string|int)[]`
+- `MemoizedTypeFactory` allows cache results of calls to `supports()` and `make()`.
+  Recommend wrap type factory with memoized decorator in production.
+### Fixed
+- Fixed string representation of collection type with complex value type.
+
 ## [1.2.0] - 2020-10-13
 ### Removed
 - `Type` interface doesn't declare static methods `support()` and `create()` no more.